requirements.txt
appdirs==1.4.3
numpy==1.12.0
packaging==16.8
pandas==0.19.2
progressbar2==3.15.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-utils==2.0.1
pytz==2016.10
six==1.10.0
